requirements.txt
absl-py==2.1.0
ansible-compat==24.5.1
ansible-core==2.15.12
ansible-lint==6.22.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.2.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
bracex==2.4
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.4.0
cryptography==42.0.7
dill==0.3.8
Faker==25.9.1
filelock==3.14.0
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.74.0
grpcio-tools==1.74.0
h5py==3.11.0
idna==3.7
importlib-resources==5.0.7
importlib_metadata==7.2.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.7
ml-dtypes==0.3.2
mypy-extensions==1.0.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
paho-mqtt==1.6.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pika==1.3.2
platformdirs==4.2.2
plumbum==1.8.3
protobuf==6.31.1
psycopg2-binary==2.9.9
py4j==0.10.9.7
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pylint==3.2.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.32.3
resolvelib==1.0.1
rich==13.7.1
rpds-py==0.18.1
rpyc==6.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
subprocess-tee==0.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcmatch==8.5.2
Werkzeug==3.0.3
wrapt==1.16.0
yamllint==1.35.1
zipp==3.19.2
